+++ libs/config/client/jest.config.ts
@@ -1,15 +1,16 @@
-module.exports = {
-  displayName: 'hapify-create-traxion-workspace',
+/* eslint-disable */
+export default {
+  displayName: 'client-config',
   preset: '../../../jest.preset.js',
   globals: {
     'ts-jest': {
       tsconfig: '<rootDir>/tsconfig.spec.json',
     },
   },
-  testEnvironment: 'node',
+  testEnvironment: 'jsdom',
   transform: {
     '^.+\\.[tj]sx?$': 'ts-jest',
   },
   moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
-  coverageDirectory: '../../../coverage/libs/hapify/create-traxion-workspace',
+  coverageDirectory: '../../../coverage/libs/config/client',
 };