+++ libs/schematics/src/generators/hapify-library/generator.spec.ts
@@ -1,5 +1,11 @@
+import { relative } from 'path';
+
 import * as angular from '@nrwl/angular/generators';
-import { readProjectConfiguration, Tree } from '@nrwl/devkit';
+import {
+  getWorkspaceLayout,
+  readProjectConfiguration,
+  Tree,
+} from '@nrwl/devkit';
 import { createTreeWithEmptyWorkspace } from '@nrwl/devkit/testing';
 import * as nestjs from '@nrwl/nest';
 import fetch, { Response } from 'node-fetch';
@@ -69,13 +75,13 @@ describe('hapify library generator', () => {
 
   it('should run successfully', async () => {
     await generator(appTree, options as HapifyLibraryGeneratorOptionsWithExtra);
-    const config = readProjectConfiguration(appTree, 'test');
+    const config = readProjectConfiguration(appTree, 'common-test');
     expect(config).toBeDefined();
   });
 
   it('should have called angular generator', async () => {
     await generator(appTree, options as HapifyLibraryGeneratorOptionsWithExtra);
-    const config = readProjectConfiguration(appTree, 'test');
+    const config = readProjectConfiguration(appTree, 'common-test');
     expect(config).toBeDefined();
     expect(angularGenerator).toHaveBeenCalledTimes(1);
     expect(nestjsGenerator).toHaveBeenCalledTimes(0);
@@ -99,13 +105,15 @@ describe('hapify library generator', () => {
   it('should generate a gitignore file', async () => {
     await generator(appTree, options as HapifyLibraryGeneratorOptionsWithExtra);
 
-    expect(appTree.exists('libs/test/.gitignore')).toBeTruthy();
-    expect(appTree.read('libs/test/.gitignore')?.toString()).toContain(
-      'hapify.json',
-    );
-    expect(appTree.read('libs/test/.gitignore')?.toString()).toContain(
-      'generated',
-    );
+    const { libsDir } = getWorkspaceLayout(appTree);
+
+    expect(appTree.exists(`${libsDir}/common/test/.gitignore`)).toBeTruthy();
+    expect(
+      appTree.read(`${libsDir}/common/test/.gitignore`)?.toString(),
+    ).toContain('hapify.json');
+    expect(
+      appTree.read(`${libsDir}/common/test/.gitignore`)?.toString(),
+    ).toContain('generated');
   });
 
   it('should generate a .hapifyrc.js file', async () => {
@@ -113,9 +121,13 @@ describe('hapify library generator', () => {
       ...options,
       hapifyTemplates: [],
     });
+    const { libsDir } = getWorkspaceLayout(appTree);
 
-    expect(appTree.exists('libs/test/.hapifyrc.js')).toBeTruthy();
-    expect(appTree.read('libs/test/.hapifyrc.js')?.toString()).toEqual(`const {
+    expect(appTree.exists(`${libsDir}/common/test/.hapifyrc.js`)).toBeTruthy();
+
+    const relativePath = relative(`${libsDir}/common/test`, '.');
+    expect(appTree.read(`${libsDir}/common/test/.hapifyrc.js`)?.toString())
+      .toEqual(`const {
   hapifyDefaultConfig,
   getValidatorPath,
 } = require('@trxn/hapify-config');
@@ -126,12 +138,12 @@ module.exports = {
   name: 'test',
   description: 'Library to host generated codes',
   validatorPath: getValidatorPath(__dirname),
-  project: '../../hapify-models.json',
+  project: '${relativePath}/hapify-models.json',
   extends: [
     '@trxn/hapify-templates-models',
   ],
   importReplacements: {
-  "mock": "@proj/models/mock"
+  "mock": "@proj/common-models/mock"
 }
 };
 `);
@@ -145,8 +157,13 @@ module.exports = {
         '@trxn/additional-templates-1,@trxn/additional-templates-2',
     });
 
-    expect(appTree.exists('libs/test/.hapifyrc.js')).toBeTruthy();
-    expect(appTree.read('libs/test/.hapifyrc.js')?.toString()).toEqual(`const {
+    const { libsDir } = getWorkspaceLayout(appTree);
+
+    expect(appTree.exists(`${libsDir}/common/test/.hapifyrc.js`)).toBeTruthy();
+
+    const relativePath = relative(`${libsDir}/common/test`, '.');
+    expect(appTree.read(`${libsDir}/common/test/.hapifyrc.js`)?.toString())
+      .toEqual(`const {
   hapifyDefaultConfig,
   getValidatorPath,
 } = require('@trxn/hapify-config');
@@ -157,12 +174,12 @@ module.exports = {
   name: 'test',
   description: 'Library to host generated codes',
   validatorPath: getValidatorPath(__dirname),
-  project: '../../hapify-models.json',
+  project: '${relativePath}/hapify-models.json',
   extends: [
     '@trxn/hapify-templates-models',
   ],
   importReplacements: {
-  "mock": "@proj/models/mock"
+  "mock": "@proj/common-models/mock"
 }
 };
 `);
@@ -174,8 +191,13 @@ module.exports = {
       hapifyTemplates: ['rext-client'],
     });
 
-    expect(appTree.exists('libs/test/.hapifyrc.js')).toBeTruthy();
-    expect(appTree.read('libs/test/.hapifyrc.js')?.toString()).toEqual(`const {
+    const { libsDir } = getWorkspaceLayout(appTree);
+
+    expect(appTree.exists(`${libsDir}/common/test/.hapifyrc.js`)).toBeTruthy();
+
+    const relativePath = relative(`${libsDir}/common/test`, '.');
+    expect(appTree.read(`${libsDir}/common/test/.hapifyrc.js`)?.toString())
+      .toEqual(`const {
   hapifyDefaultConfig,
   getValidatorPath,
 } = require('@trxn/hapify-config');
@@ -186,12 +208,12 @@ module.exports = {
   name: 'test',
   description: 'Library to host generated codes',
   validatorPath: getValidatorPath(__dirname),
-  project: '../../hapify-models.json',
+  project: '${relativePath}/hapify-models.json',
   extends: [
     '@trxn/hapify-templates-models',
   ],
   importReplacements: {
-  "mock": "@proj/models/mock"
+  "mock": "@proj/common-models/mock"
 }
 };
 `);
@@ -200,7 +222,10 @@ module.exports = {
   it('should add the generate target to the project configuration', async () => {
     await generator(appTree, options as HapifyLibraryGeneratorOptionsWithExtra);
 
-    const projectConfiguration = readProjectConfiguration(appTree, 'test');
+    const projectConfiguration = readProjectConfiguration(
+      appTree,
+      'common-test',
+    );
 
     expect(targetGenerator).toBeCalledTimes(1);
     expect(projectConfiguration.targets?.generate).toBeDefined();