+++ libs/schematics/src/generators/traxion-workspace/helpers/create-angular-application.helper.ts
@@ -1,5 +1,6 @@
 import { Tree } from '@nrwl/devkit';
 
+import { getLogger } from '../../../helpers';
 import pwaAppGenerator from '../../pwa-app/generator';
 import { NormalizedOptions } from '../schema';
 
@@ -7,11 +8,14 @@ export async function createAngularApplication(
   tree: Tree,
   options: NormalizedOptions,
 ) {
-  const { pwaName, generatedDir, apiName, directory } = options;
-  await pwaAppGenerator(tree, {
-    name: pwaName,
-    directory,
-    generatedDir,
-    apiName,
-  });
+  const log = getLogger(tree);
+  log.warn('Pwa application schematics is disabled for now');
+
+  // const { pwaName, generatedDir, apiName, directory } = options;
+  // await pwaAppGenerator(tree, {
+  //   name: pwaName,
+  //   directory,
+  //   generatedDir,
+  //   apiName,
+  // });
 }