+++ libs/schematics/src/generators/traxion-workspace/helpers/create-nestjs-application.helper.ts
@@ -1,5 +1,6 @@
 import { Tree } from '@nrwl/devkit';
 
+import { getLogger } from '../../../helpers';
 import apiAppGenerator from '../../api-app/generator';
 import { NormalizedOptions } from '../schema';
 
@@ -7,10 +8,13 @@ export async function createNestjsApplication(
   tree: Tree,
   options: NormalizedOptions,
 ) {
-  const { directory, apiName, generatedDir } = options;
-  await apiAppGenerator(tree, {
-    name: apiName,
-    directory,
-    generatedDir,
-  });
+  const log = getLogger(tree);
+  log.warn('Nestjs application schematics is disabled for now');
+
+  // const { directory, apiName, generatedDir } = options;
+  // await apiAppGenerator(tree, {
+  //   name: apiName,
+  //   directory,
+  //   generatedDir,
+  // });
 }